document-iq/document-iq-platform/components/application-component/pyproject.toml
[project]
name = "document-iq-platform-application"
version = "1.0.1"
description = "Document aggregator service for DocumentIQ platform"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "kafka-python",
    "pydantic",
    "python-dotenv",
    "redis",
    "sqlalchemy",
    "alembic",
    "passlib[bcrypt]>=1.7.4,<1.8",
    "bcrypt>=4.0.1,<4.1",
    "python-jose",
    "pymysql",
    "pydantic[email]",
    "document-iq-platform-shared"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]